+++ packages/design-system/src/atoms/buttonV2/index.tsx
@@ -411,3 +411,5 @@ export const ButtonV2 = ({
     />
   )
 }
+
+export * from './types'
