+++ packages/design-system/src/atoms/index.tsx
@@ -17,10 +17,10 @@ export * from './input'
 export * from './icon'
 export * from './inputV2ActionSuffix'
 export * from './inputV2TextSuffix'
-export * from './inputV2SelectSuffix'
 export * from './toastV2'
 export * from './tooltipV2'
 export * from './statusV2'
 export * from './switch'
 export * from './radioSelect'
 export * from './qrcode'
+export * from './tokenSelectOptionV2'
