+++ packages/design-system/src/hooks/index.tsx
@@ -1,4 +1,3 @@
-export * from './web3Wagmi'
 export * from './theme'
 export * from './themeV2'
 export * from './web3'
@@ -6,3 +5,4 @@ export * from './modal'
 export * from './native'
 export * from './clipboard'
 export * from './breakpoint'
+export * from './web3Wagmi'
