+++ packages/design-system/src/molecules/index.ts
@@ -31,3 +31,4 @@ export * from './web3Button'
 export * from './estimatedGas'
 export * from './inputContainer'
 export * from './accountInfoDropdown'
+export * from './inputV2SelectSuffix'
