+++ packages/design-system/src/molecules/inputV2SelectSuffix/index.tsx
@@ -2,10 +2,10 @@ import { Platform, View } from 'react-native'
 import { useCallback, useRef } from 'react'
 import { Pressable } from '@gluestack-ui/react'
 import { useThemeV2 } from '../../hooks'
-import { TypographyV2 } from '../typographyV2'
+import { TypographyV2 } from '../../atoms/typographyV2'
 import { container, spacing } from '../../styles'
-import { Icon } from '../icon'
-import { SelectV2, SelectV2Props } from '../../molecules/selectV2'
+import { Icon } from '../../atoms/icon'
+import { SelectV2, SelectV2Props } from '../selectV2'
 
 export type InputV2SelectSuffixProps = Omit<SelectV2Props, 'renderReference'>
