+++ packages/design-system/src/molecules/sidebar/mobile.tsx
@@ -10,9 +10,7 @@ import {
 } from 'react-native'
 import { useSafeAreaInsets } from 'react-native-safe-area-context'
 import { AnimatePresence, Motion } from '@legendapp/motion'
-import { Icon } from '../../atoms'
-import { ButtonV2 } from '../../atoms/buttonV2'
-import { TypographyV2 } from '../../atoms/typographyV2'
+import { TypographyV2, Icon, ButtonV2 } from '../../atoms'
 import { useThemeV2 } from '../../hooks'
 import { container, layer, margin, padding } from '../../styles'
 import { openLink } from '../../utils'
