+++ packages/design-system/src/organisms/formStepHeader/index.tsx
@@ -1,6 +1,6 @@
 import { useCallback, useMemo, useRef } from 'react'
 import { LayoutChangeEvent, Platform, StyleSheet, View } from 'react-native'
-import { TypographyV2 } from '../../atoms/typographyV2'
+import { TypographyV2 } from '../../atoms'
 import { useThemeV2, useBreakpointValue } from '../../hooks'
 import { container, margin, spacing } from '../../styles'
